ref: refactor code structure
